+++ src/lib/components/delegations/DelegationInfo.tsx
@@ -33,7 +33,10 @@ export const DelegationInfo = ({
           bg="gray.900"
           borderRadius="8px"
           p={4}
-          onClick={onViewMore}
+          onClick={() => {
+            trackUseViewMore();
+            onViewMore();
+          }}
         >
           <Flex direction="column" gap={2}>
             <Heading variant="h6" as="h6">