+++ src/lib/pages/proposal-details/components/proposal-top/index.tsx
@@ -101,7 +101,7 @@ export const ProposalTop = ({ proposalData }: ProposalTopProps) => {
                             fontWeight: 600,
                           }}
                         >
-                          {" / "}
+                          {" , "}
                         </span>
                       )}
                       {msgType}