+++ src/lib/pages/validator-details/index.tsx
@@ -13,6 +13,7 @@ import { Loading } from "lib/components/Loading";
 import PageContainer from "lib/components/PageContainer";
 import PageHeaderContainer from "lib/components/PageHeaderContainer";
 import { ErrorFetching, InvalidState } from "lib/components/state";
+import { UserDocsLink } from "lib/components/UserDocsLink";
 import { useAssetInfos } from "lib/services/assetService";
 import { useValidatorData } from "lib/services/validatorService";
 
@@ -135,6 +136,11 @@ const ValidatorDetailsBody = ({
             </TabPanel>
           </TabPanels>
         </Tabs>
+        <UserDocsLink
+          title="What is a Validator?"
+          cta="Read more about Validator Details"
+          href="general/validator/detail-page"
+        />
       </PageContainer>
     </>
   );