+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- [#878](https://github.com/alleslabs/celatone-frontend/pull/878) Support searching proposal id and validator address on the main search
 - [#875](https://github.com/alleslabs/celatone-frontend/pull/875) Add Amplitude tracking to validator details page
 - [#865](https://github.com/alleslabs/celatone-frontend/pull/865) Apply voted proposals filter and search input to table
 - [#860](https://github.com/alleslabs/celatone-frontend/pull/860) Add voted proposals in voted tab
@@ -133,6 +134,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Bug fixes
 
 - [#877](https://github.com/alleslabs/celatone-frontend/pull/877) Fix validator list/details bug bash
+- [#874](https://github.com/alleslabs/celatone-frontend/pull/874) Fix minor proposal and validator bugs
+- [#876](https://github.com/alleslabs/celatone-frontend/pull/876) Fix validator votes dropdown
 - [#873](https://github.com/alleslabs/celatone-frontend/pull/873) Fix bug bash select input and ui
 - [#868](https://github.com/alleslabs/celatone-frontend/pull/868) Fix validator jail state
 - [#867](https://github.com/alleslabs/celatone-frontend/pull/867) Fix validator details voting table